implement animation for Java. Type conversion hell!

+++ src/com/jdrago/blackout/BlackoutRenderer.java
@@ -24,7 +24,7 @@ public class BlackoutRenderer extends QuadRenderer
         textures_.put("cards", loadPNG(R.raw.cards));
     }
 
-    public void blit(String textureName, int srcX, int srcY, int srcW, int srcH, int dstX, int dstY, int dstW, int dstH, float rot, float anchorX, float anchorY)
+    public void blit(String textureName, float srcX, float srcY, float srcW, float srcH, float dstX, float dstY, float dstW, float dstH, float rot, float anchorX, float anchorY)
     {
         blit(textures_.get(textureName), srcX, srcY, srcW, srcH, dstX, dstY, dstW, dstH, rot, anchorX, anchorY);
     }
